% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_conversion_factors.R
\name{compute_conversion_factors}
\alias{compute_conversion_factors}
\title{Conversion factors computation}
\usage{
compute_conversion_factors(m, max_dt = NULL, time1 = 0, time2 = 1)
}
\arguments{
\item{m}{Jsmm named list that contains the fitted model.}

\item{max_dt}{positive real number that describes the maximal time resolution used in numerical computations.}

\item{time1}{positive real number that defines the starting time for the time duration period T.}

\item{time2}{positive real number that defines the end time for the time duration period T.}
}
\value{
A matrix with the computations for the species defined in m.
}
\description{
This function implements the conversion factor \eqn{Z^* = \frac{1}{Ap}} method described in Rodriguez, Gårdman, et al. (2026).  \eqn{Z^*} represents the population density that results in the expected number of trapped individuals being \eqn{E[N] = 1}. \eqn{A} and \eqn{p} represent the domain area and probability by which the individual would become trapped according to the continuous capture process, respectively.
}
\references{
Rodriguez L.F., Gårdman V., Roslin T., & Ovaskainen O. (2026). Malaise trap samples of 1000 individuals per week suggest 4 million insects per hectare in the boreal zone.
}
