% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_likelihood.R
\name{compute_likelihood}
\alias{compute_likelihood}
\title{Compute the likelihood}
\usage{
compute_likelihood(m, Beta, max_dt = NULL)
}
\arguments{
\item{m}{Named list that contains all the model components for running Joint species movement modelling (jsmm) analyses.}

\item{Beta}{A named list with the movement and observation parameters included in the Jsmm model.}

\item{max_dt}{Optional positive integer value representing the number of time partitions in which the PDE should be solved.}
}
\value{
A named list with the loglikelihood computations for each species.
}
\description{
Compute the likelihood in log form for the Joint species movement model
}
