% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_coda_object.R
\name{convert_to_coda_object}
\alias{convert_to_coda_object}
\title{Convert to Coda object}
\usage{
convert_to_coda_object(
  m,
  start = 1,
  Beta = TRUE,
  Gamma = TRUE,
  rho = TRUE,
  SI = TRUE,
  li = TRUE
)
}
\arguments{
\item{m}{Named list that contains all the model components for running Joint species movement modelling (jsmm) analyses.}

\item{start}{Start of the MCMC}

\item{Beta}{Optional logical data type. Set to TRUE to include Beta parameters, FALSE otherwise.}

\item{Gamma}{Optional logical data type. Set to TRUE to include Gamma parameters, FALSE otherwise.}

\item{rho}{Optional logical data type. Set to TRUE to include rho parameters, FALSE otherwise.}

\item{SI}{Optional logical data type. Set to TRUE to include Sigma parameters, FALSE otherwise.}

\item{li}{Optional logical data type. Set to TRUE to include Markov chain Monte Carlo chains, FALSE otherwise.}
}
\value{
A named list with objects from the class mcmc from coda.
}
\description{
Converts the posterior data generated after fitting the jsmm model into an object to the class mcmc from the Coda R-package.
}
