% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creating_domain_from_locations.R
\name{creating_domain_from_locations}
\alias{creating_domain_from_locations}
\title{Create a simulated domain}
\usage{
creating_domain_from_locations(
  lc,
  lc_label = NA,
  buffer_x = c(0, 0),
  buffer_y = c(0, 0),
  nnod_side_x,
  nnod_side_y,
  rad_capture_site,
  n_sides_capture_site,
  angle_capture_site,
  int_buffer = FALSE
)
}
\arguments{
\item{lc}{matrix that provides the positions of the capture and release sites. Columns correspond to the x and y coordinates, respectively.}

\item{lc_label}{character vector with capture/release sites labels (optional).}

\item{buffer_x}{real positive number representing the distance from the center of the outer capture/release location to the domain boundary in the x-axis.}

\item{buffer_y}{real positive number representing the distance from the center of the outer capture/release location to the domain boundary in the y-axis.}

\item{nnod_side_x}{integer positive representing the number of nodes in x axes.}

\item{nnod_side_y}{integer positive representing the number of nodes in y axes.}

\item{rad_capture_site}{real positive value representing the capture/release site radius.}

\item{n_sides_capture_site}{Number of sides for the capture/release polygon.}

\item{angle_capture_site}{input parameter in radians measuring the counterclockwise rotation angle for the capture/release polygon.}

\item{int_buffer}{boolean parameter for inserting an intermediate buffer between the border of the captures/releases array and the landscape boundary.}
}
\value{
Polygonal object from the class sf.
}
\description{
This function allows the creation of a customized rectangular domain, capture, and release sites.
}
