% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsmm.R
\name{jsmm}
\alias{jsmm}
\title{Define the Jsmm model}
\usage{
jsmm(
  domain,
  observation_effort,
  releases = NULL,
  captures = NULL,
  cov_data,
  ns,
  sp_names = NULL,
  C = NULL,
  tr_data = NULL,
  model_formula
)
}
\arguments{
\item{domain}{named list with the elements domain and triangulation. The former is a polygonal object from the class sf, and the latter provides information about the domain triangulation.}

\item{observation_effort}{named list with spatial and temporal information about releases and captures in the experiment.}

\item{releases}{named list with information about where and when releases occurred.}

\item{captures}{named list with information about where and when captures occurred.}

\item{cov_data}{named list with (some combination of) spatial, temporal or/and spatio-temporal data.}

\item{ns}{integer representing the number of species included in the model.}

\item{sp_names}{string vector parameter assigning the species names in the model (optional).}

\item{C}{phylogenetic relationships matrix between species.}

\item{tr_data}{dataframe with trait data.}

\item{model_formula}{named list that consists of (some combination of) the objects diffusion, advection, mortality, habitat_preference, observation, and traits. These follow the R-notation for the model formulae.}
}
\value{
A named list with all the information for running Jsmm analyses.
}
\description{
Creates a named list with all the necessary components for running the Jsmm analyses.
}
