% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsmm_add_triangulation.R
\name{jsmm_add_triangulation}
\alias{jsmm_add_triangulation}
\title{Add triangulation}
\usage{
jsmm_add_triangulation(domain, max_t_area, min_t_angle)
}
\arguments{
\item{domain}{polygonal object from the class sf.}

\item{max_t_area}{Maximum triangle area.}

\item{min_t_angle}{Minimum triangle angle in degrees.}
}
\value{
A named list with the triangulation and its corresponding polygonal sf object.
}
\description{
Create a domain triangulation from a polygonal object from the class sf.
}
