% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_captures.R
\name{mean_captures}
\alias{mean_captures}
\title{Compute the captured individuals mean}
\usage{
mean_captures(m, captures)
}
\arguments{
\item{m}{Named list that contains all the model components for running Joint species movement modelling (Jsmm) analyses.}

\item{captures}{Matrix with the capture data.}
}
\value{
A vector with the estimates for each species.
}
\description{
This function computes the mean of the captured individual for each species. This function is used for evaluating the model fit (posterior predictive data).
}
