% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_distance_between_captures.R
\name{mean_distance_between_captures}
\alias{mean_distance_between_captures}
\title{Compute the mean distance between captures}
\usage{
mean_distance_between_captures(m, captures)
}
\arguments{
\item{m}{Named list that contains all the model components for running Joint species movement modelling (Jsmm) analyses.}

\item{captures}{Matrix with the capture data.}
}
\value{
A vector with the computations for each species.
}
\description{
This function computes the mean euclidean distance between captures for each species. This function is used for evaluating the model fit (posterior predictive data).
}
