% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_time_to_capture.R
\name{mean_time_to_capture}
\alias{mean_time_to_capture}
\title{Compute the mean time between captures}
\usage{
mean_time_to_capture(m, captures)
}
\arguments{
\item{m}{Named list that contains all the model components for running Joint species movement modelling (Jsmm) analyses.}

\item{captures}{Matrix with the capture data.}
}
\value{
A vector with the computations for each species.
}
\description{
This function computes the mean time between captures for each species. This function is used for evaluating the model fit (posterior predictive data).
}
