% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_rc_data.R
\name{merge_rc_data}
\alias{merge_rc_data}
\title{Merge releases and capture data}
\usage{
merge_rc_data(m)
}
\arguments{
\item{m}{Named list that contains all the model components for running Joint species movement modelling (Jsmm) analyses.}
}
\value{
A modified Jsmm model m.
}
\description{
Sets releases and capture data.
}
