% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_summary.R
\name{model_summary}
\alias{model_summary}
\title{Show summary of the Jsmm model.}
\usage{
model_summary(m)
}
\arguments{
\item{m}{named list corresponding to the Jsmm model.}
}
\value{
Produces a visualization on screen of the Jsmm main components.
}
\description{
Generates a summary of all the Jsmm components.
}
