% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_of_captures.R
\name{number_of_captures}
\alias{number_of_captures}
\title{Calculate the number of captures}
\usage{
number_of_captures(m, captures)
}
\arguments{
\item{m}{Named list that contains all the model components for running Joint species movement modelling (Jsmm) analyses.}

\item{captures}{Matrix with the capture data.}
}
\value{
A vector with the computations for each species.
}
\description{
This function computes captured individuals for each species. This function is used for evaluating the model fit (posterior predictive data).
}
