% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_domain.R
\name{plot_domain}
\alias{plot_domain}
\title{Plot domain}
\usage{
plot_domain(domain, customize_plot = NULL)
}
\arguments{
\item{domain}{polygonal object from the class sf.}

\item{customize_plot}{list with optional presetting for visualization.}
}
\value{
A customized domain plot.
}
\description{
Create a plot from the polygonal triangulation.
}
