% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_observation_effort.R
\name{plot_observation_effort}
\alias{plot_observation_effort}
\title{Plot Observation effort}
\usage{
plot_observation_effort(
  observation_effort,
  xlab,
  ylab,
  by_x,
  cex.axisx,
  cex.axisy,
  lasy,
  tcky
)
}
\arguments{
\item{observation_effort}{Named list with releases and captures information.}

\item{xlab}{Customized labels for the x-axis.}

\item{ylab}{Customized labels for the y-axis.}

\item{by_x}{Ticks in x frequency.}

\item{cex.axisx}{Size text x.}

\item{cex.axisy}{Size text y.}

\item{lasy}{Rotation label y.}

\item{tcky}{Control tick marks.}
}
\value{
A plot with releases and captures information.
}
\description{
Visualize where and when captures and releases occurred.
}
