% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trajectories.R
\name{plot_trajectories}
\alias{plot_trajectories}
\title{Plot movement trajectories.}
\usage{
plot_trajectories(m = m, plot_sp = NULL, customize_plot = NULL)
}
\arguments{
\item{m}{Named list that contains all the model components for running Joint species movement modelling (Jsmm) analyses.}

\item{plot_sp}{vector composed of integers representing the species index to visualize.}

\item{customize_plot}{list with predefined settings for visualization.}
}
\value{
NULL. Visualization function.
}
\description{
The function allows visualization of the movement trajectories generated by the function simulate_trajectories().
}
