% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poolMcmcChains.R
\name{poolMcmcChains}
\alias{poolMcmcChains}
\title{Pool MCMC chains}
\usage{
poolMcmcChains(postList, chainIndex = 1:length(postList), start = 1, thin = 1)
}
\arguments{
\item{postList}{List with the model posterior data}

\item{chainIndex}{Index of the MCMC chain.}

\item{start}{Starting point MCMC chain.}

\item{thin}{Thin MCMC chain}
}
\value{
A list with the formatted posterior data.
}
\description{
This function sets up the posterior data in suitable format. Used for showing posterior vs priors.
}
