% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleMcmc.R
\name{sampleMcmc}
\alias{sampleMcmc}
\title{Sample MCMC}
\usage{
sampleMcmc(
  m,
  samples,
  transient,
  thin,
  max_dt = NULL,
  init_pars = NULL,
  nChains = 1,
  no.data = FALSE
)
}
\arguments{
\item{m}{Named list that contains all the model components for running Joint species movement modelling (Jsmm) analyses.}

\item{samples}{Integer parameter representing the number of samples drawn from the MArkov Chain.}

\item{transient}{Integer parameter representing the initial part of the chain  considered as transient.}

\item{thin}{Integer parameter representing the thinning.}

\item{max_dt}{Optional positive integer value representing the number of time partitions in which the PDE should be solved.}

\item{init_pars}{Initial parameters for fitting the model.}

\item{nChains}{Number of MCMC chains to simulate.}

\item{no.data}{Boolean parameter indicating presence or absence of data.}
}
\value{
This functions adds to the named list m, the posterior data, thin, samples, and transient information.
}
\description{
Performs all the computations for the parameter estimation for the the Joint species movement model (Jsmm) via Markov Chain Monte Carlo (MCMC).
}
