% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_captures.R
\name{simulate_captures}
\alias{simulate_captures}
\title{Simulate movement data.}
\usage{
simulate_captures(m, pars, releases, max_dt = NULL, secondary_release)
}
\arguments{
\item{m}{named list created by the function jsmm().}

\item{pars}{named list with the objects Beta, rho, SI, and Gamma defining the model parameters.}

\item{releases}{matrix whose first and second columns provide information about the species and release event, respectively.}

\item{max_dt}{real number bigger than zero that describes the maximal time resolution used in numerical computations.}

\item{secondary_release}{boolean parameter that describes whether the captured individuals are released after capture.}
}
\value{
The function returns the model object m, to which it has added the objects captures and releases.
}
\description{
The function returns the model object m, to which it has added the objects captures and releases.
}
