% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_from_prior.R
\name{simulate_from_prior}
\alias{simulate_from_prior}
\title{Simulate movement and observation parameters from priors.}
\usage{
simulate_from_prior(m)
}
\arguments{
\item{m}{Named list that contains all the model components for running Joint species movement modelling (Jsmm) analyses.}
}
\value{
A named list with the movement and observation parameters, and the corresponding rho, Sigma, and Gamma priors.
}
\description{
This function generates the movement and observation parameters from the priors rho (phylogenetic signal parameter), parameter Sigma for generating the variance-covariance matrix, and vector Gamma.
}
