% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_trajectories.R
\name{simulate_trajectories}
\alias{simulate_trajectories}
\title{Simulate movement trajectories.}
\usage{
simulate_trajectories(m, pars, releases, max_dt = NULL)
}
\arguments{
\item{m}{named list created by the function jsmm().}

\item{pars}{named list with the objects Beta, rho, SI, and Gamma defining the model parameters.}

\item{releases}{matrix whose first and second columns provide information about the species and release event, respectively.}

\item{max_dt}{real number bigger than zero that describes the maximal time resolution used in numerical computations.}
}
\value{
#' The function returns the model object m, to which it has been added a matrix called trajectories.
}
\description{
The function generates trajectories for a movement model in which secondary releases are not possible.
}
